LayoutTests/imported/w3c/web-platform-tests/screen-orientation/lock-sandboxed-iframe-expected.txt

CONSOLE MESSAGE: Error while parsing the 'sandbox' attribute: 'allow-orientation-lock' is an invalid sandbox flag.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: document.documentElement.requestFullscreen is not a function. (In 'document.documentElement.requestFullscreen()', 'document.documentElement.requestFullscreen' is undefined)


Harness Error (TIMEOUT), message = null

TIMEOUT Test without 'allow-orientation-lock' sandboxing directive Test timed out
NOTRUN Test with 'allow-orientation-lock' sandboxing directive
